Tweak wording of flow control
We need to be very clear which component of the receiver is being
engaged.  The QUIC transport has to process the packet if it
acknowledges it, but that doesn't mean processing the contents of the
DATAGRAM frame.

(I also killed a bad character that wasn't a space but should have been
in the next section.)